+++ NEWS.md
@@ -4,3 +4,4 @@
 * Added `constantEggRDI()` to allow keeping egg densities fixed.
 * Vignette explaining size-spectrum dynamics in single-species example.
 * New function `customFunction()` to allow users to overwrite mizer functions.
+* New function `projectToSteady()` to run the full dynamics to steady state.